Change to signed-16bit wavetable.

+++ user/WaveTable_Celesta_C5.h
@@ -1,9 +1,11 @@
 #ifndef __WAVETABLE_CELESTA_C5__
 #define __WAVETABLE_CELESTA_C5__
 #include <stdint.h>
+// Sample's base frequency: 523.250000 Hz
+// Sample's sample rate: 32000 Hz
 #define WAVETABLE_CELESTA_C5_LEN 2608
 #define WAVETABLE_CELESTA_C5_ATTACK_LEN 1998
 #define WAVETABLE_CELESTA_C5_LOOP_LEN 610
-extern const int8_t WaveTable_Celesta_C5[WAVETABLE_CELESTA_C5_LEN];
+extern const int16_t WaveTable_Celesta_C5[WAVETABLE_CELESTA_C5_LEN];
 extern const uint16_t WaveTable_Celesta_C5_Increment[];
 #endif
